--- src/components/MapInterface/ShowOSMNotes/index.tsx
@@ -1,1 +1,0 @@
-export * from './ShowOSMNotes'
