+++ src/components/MapInterface/ShowOsmNotes/index.tsx
@@ -1,0 +1,1 @@
+export * from './ShowOsmNotes'
